requirements.txt
jsonschema
pyld
rfc3987
jsonref
rdflib-jsonld
pyOpenSSL
notebook==5.0.0